refactor: handler 로직 분리
